backend/requirements.txt
fastapi==0.109.0
uvicorn>=0.30.0,<0.32
sqlalchemy==2.0.38
pydantic==2.12.5
python-dotenv==1.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-telegram-bot>=21.0,<22
httpx>=0.26.0
socksio>=1.0.0
redis>=5.0.0,<6
jdatetime==4.1.1
openpyxl==3.1.5